Sec.11: use react-helmet to inject css font
Please do not use @import CSS syntax in createGlobalStyle
at this time, as the CSSOM APIs we use in production do not
handle it well.
Instead, we recommend using a library such as react-helmet
to inject a typical <link> meta tag to the stylesheet, or simply
embedding it manually in your index.html <head> section for
a simpler app.